Made renderer.begin take a Projection union, moved vsync to a runtime option, switched clear to u8, snapped sprites to whole pixels, and ported UiContext to the public renderer
- renderer.begin(.logical | .screen | .{ .camera = &cam } | .{ .matrix = m });
  the renderer holds the engine's viewport to resolve them.
- Renderer gained flush(), setClip(?RectF), drawClippedString and lineHeight.
- imgui.UiContext(Engine).init(eng) draws through the renderer's public
  methods; its manual flushes and scissor code are gone. Console.draw takes
  the generic ui as anytype.
- PixzigEngineOptions.vsyncEnabled became PixzigEngineInitOptions.vsync.
- renderer.clear and pz_render_clear take 0-255 u8 values.
- SpriteBatchQueue.drawSprite rounds the sprite's top-left to a whole pixel.
- grid_render_ex uses eng.projection(); create_texture_ex was rewritten on
  AppRunner and added to the build.

+++ python/pixzig/_native.py
@@ -129,7 +129,7 @@ pz_finish_tick = _sig("pz_finish_tick", [PzEnginePtr], None)
 pz_swap_buffers = _sig("pz_swap_buffers", [PzEnginePtr], None)
 pz_render_begin = _sig("pz_render_begin", [PzEnginePtr], None)
 pz_render_begin_world = _sig("pz_render_begin_world", [PzEnginePtr, PzCameraPtr], None)
-pz_render_clear = _sig("pz_render_clear", [PzEnginePtr, c_float, c_float, c_float, c_float], None)
+pz_render_clear = _sig("pz_render_clear", [PzEnginePtr, c_uint8, c_uint8, c_uint8, c_uint8], None)
 pz_render_end = _sig("pz_render_end", [PzEnginePtr], None)
 
 # --- Input ---------------------------------------------------------------
